ref: refactor router api to separate package
